requirements.txt
Flask
Flask-CORS
langchain
langchain-groq
langchain-community
python-dotenv